# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reviews/{id}": {
            "get": {
                "tags": [
                    "View-review-for-specific-property"
                ],
                "summary": "Get reviews for a specific property",
                "description": "Fetch all reviews for a given property.",
                "operationId": "6eeda9b36bfd05a475a55a79473890f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the property to fetch reviews for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "property_id": {
                                            "type": "integer"
                                        },
                                        "reviews": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Review"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reviews/store": {
            "post": {
                "tags": [
                    "User-write-review"
                ],
                "summary": "Store a new review for a property",
                "description": "Allow users to add a review for a specific property.",
                "operationId": "405dd3898832581226182c80f4a7ff5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "property_id",
                                    "comment",
                                    "review"
                                ],
                                "properties": {
                                    "property_id": {
                                        "type": "integer"
                                    },
                                    "comment": {
                                        "type": "string"
                                    },
                                    "review": {
                                        "description": "Rating between 1 and 5",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully added review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "review": {
                                            "$ref": "#/components/schemas/Review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reviews/delete/{reviewId}": {
            "delete": {
                "tags": [
                    "Delete-written-review"
                ],
                "summary": "Delete a review by review ID (admin or general delete)",
                "description": "Delete any review by its review ID (admin or moderator functionality).",
                "operationId": "2682b63a24f1558d54b323820c5b4efe",
                "parameters": [
                    {
                        "name": "reviewId",
                        "in": "path",
                        "description": "The ID of the review to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Review not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/property/{id}/reviews": {
            "get": {
                "tags": [
                    "View-property-review"
                ],
                "summary": "Get all reviews for a specific property",
                "description": "Fetch all reviews written for a given property by property ID.",
                "operationId": "3e3ef69d81fbb072009aaeb4ce9612a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the property to fetch reviews for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Review"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No reviews found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Review": {
                "description": "A user review for a property",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "properties_id": {
                        "type": "integer"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "Rating": {
                        "description": "Rating from 1 to 5",
                        "type": "integer"
                    },
                    "flag": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Bearer token to access protected routes",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}